Refactor navigation swipe handler in client/ package as well
